refa and added logging and DI
