+++ src/cloud/IoTalkie/Messaging/Channel/Devices/MessageController.cs
@@ -1,14 +1,11 @@
 ﻿using System;
-using System.Diagnostics;
-using System.IO;
 using System.Linq;
-using System.Net.Http.Headers;
 using System.Threading.Tasks;
+using IoTalkie.Messaging.Processing;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
-using Microsoft.Extensions.Primitives;
 
-namespace IoTalkie.Messaging
+namespace IoTalkie.Messaging.Channel.Devices
 {
     [Route("api/message")]
     [ApiController]
@@ -74,61 +71,5 @@ namespace IoTalkie.Messaging
             return this.Ok();
         }
 
-    }
-
-    public class RoutingMessage
-    {
-        private readonly string _messageId;
-        private AzureBlobPayload _payload;
-
-        public RoutingMessage(string messageId, DevicePrincipal sender, ContactPrincipal recipient, AzureBlobPayload payload)
-        {
-            Sender = sender;
-            _messageId = messageId;
-            Recipient = recipient;
-            _payload = payload;
-        }
-
-        public Principal Sender { get; set; }
-
-        public Principal Recipient { get; set; }
-
-        public AzureBlobPayload Payload => _payload;
-
-        public string MessageId => _messageId;
-    }
-
-    public class ContactPrincipal : Principal
-    {
-        private readonly string _userId;
-        private readonly Principal _original;
-
-        public ContactPrincipal(string userId, Principal original = null)
-        {
-            _userId = userId;
-            _original = original;
-        }
-
-        public string UserId => _userId;
-
-        public Principal Original => _original;
-    }
-
-    public class DevicePrincipal : Principal
-    {
-        private readonly string _clientId;
-        private readonly string _userAgent;
-
-        public DevicePrincipal(string clientId, string userAgent)
-        {
-            _clientId = clientId;
-            _userAgent = userAgent;
-        }
-
-        public string ClientId => _clientId;
-    }
-
-    public abstract class Principal
-    {
     }
 }
