+++ src/cloud/IoTalkie/Messaging/Channel/Principal.cs
@@ -1,0 +1,6 @@
+﻿namespace IoTalkie.Messaging.Channel
+{
+    public abstract class Principal
+    {
+    }
+}
